Only change available locales during org creation
