+++ decidim-admin/app/commands/decidim/admin/update_organization.rb
@@ -38,7 +38,6 @@ module Decidim
         {
           name: form.name,
           description: form.description,
-          available_locales: form.available_locales,
           default_locale: form.default_locale
         }
       end
