+++ decidim-admin/app/forms/decidim/admin/organization_form.rb
@@ -13,15 +13,19 @@ module Decidim
       attribute :current_user, Decidim::User
 
       attribute :name, String
-      attribute :available_locales, Array
       attribute :default_locale, String
       translatable_attribute :description, String
 
       validates :name, presence: true
-      validates :available_locales, presence: true
       validates :default_locale, presence: true
       validates :default_locale, inclusion: { in: :available_locales }
       validates :description, translatable_presence: true
+
+      private
+
+      def available_locales
+        current_organization.available_locales
+      end
     end
   end
 end
