+++ decidim-admin/app/views/decidim/admin/organization/_form.html.erb
@@ -6,26 +6,6 @@
   <%= form.translated :editor, :description %>
 </div>
 
-<%= form.fields_for :locales do |fields| %>
 <div class="field">
-  <%= form.label :organization_locales, '', class: @form.respond_to?(:errors) && @form.errors[:default_locale].present? ? "is-invalid-label" : "" %>
-  <table>
-    <thead>
-      <tr>
-        <td>Locale</td>
-        <td>Enabled <%= form.error_for(:available_locales) %></td>
-        <td>Default? <%= form.error_for(:default_locale) %></td>
-      </tr>
-    </thead>
-    <tbody>
-      <% localized_locales.each do |locale| %>
-        <tr>
-          <td><%= locale.name %></td>
-          <td><%= check_box_tag "organization[available_locales][#{locale.id}]", locale.id, @form.available_locales.include?(locale.id) %></td>
-          <td><%= radio_button_tag "organization[default_locale]", locale.id, @form.default_locale == locale.id %></td>
-        </tr>
-      <% end %>
-    </tbody>
-  </table>
+  <%= form.collection_select :default_locale, localized_locales(current_organization.available_locales), :id, :name %>
 </div>
-<% end %>
