+++ decidim-system/app/assets/javascripts/decidim/system/application.js.es6
@@ -2,6 +2,7 @@
 //= require jquery_ujs
 //= require foundation
 //= require turbolinks
+//= require decidim/editor
 //= require_self
 
 $(document).on('turbolinks:load', () => {
